Badly needed patch for graphics package (GWTCanvas and ImageLoader). Much Improved IE implementation. Performance improvements (via fast string appends) and pixel for pixel correctness with native canvas (thanks to Oliver Zoran for his help).
Also layed groundwork for Gradients. They work in all browsers except IE, but the framework is there for a contributor to pick up and finish.

Fixed some bugs with ImageLoader and moved over to use the ImageElement class in the new DOM api instead of the old custom image handle.

Added new demos.